[Feature] Implement countries screen with API

+++ lib/world_information/model/world_information_model.dart
@@ -9,7 +9,7 @@ class WorldInformationModel {
   int active;
   int critical;
   int casesPerOneMillion;
-  int deathsPerOneMillion;
+  double deathsPerOneMillion;
   int tests;
   double testsPerOneMillion;
   int population;
